refactor: removed dead code
