+++ src/contexts/index.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
-import { AuthProvider } from "./Auth";
-import { ToastProvider } from "./useToast";
+import { AuthProvider } from "./AuthProvider";
+import { ToastProvider } from "./ToastProvider";
 
 interface AppProviderProps {
   children: ReactNode;
